requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.30
PyMySQL==1.1.1
Jinja2==3.1.4
email-validator==2.3.0
python-multipart==0.0.20
